fix: cast mpv newtype-enum discriminants to u32 on windows
bindgen 0.72 emits `mpv_log_level`/`mpv_event_id` with an `i32` inner on
the windows clang/MinGW targets (both x86_64-pc-windows-msvc via CLANG64
and aarch64-pc-windows-gnullvm via CLANGARM64) and `u32` on Linux. The
`LogLevel` repr(u32) discriminants and `Event::Other(u32)` failed to
build on both windows CI jobs. Cast `.0 as u32` at the call sites
instead of leaning on bindgen's per-toolchain signedness choice.

+++ src/mpv/src/event.rs
@@ -220,7 +220,7 @@ impl Event {
                     name: cstr_to_string(h.name),
                 }
             },
-            other => Event::Other(other.0),
+            other => Event::Other(other.0 as u32),
         }
     }
 }
